requirements.txt
jinja2~=3.1.4
pandas~=1.3.5
numpy~=1.21
graphviz
MarkupSafe~=2.1.5
matplotlib
pandera[pandas]
graphviz
modAL-python
scikit-learn
scipy